+++ Example/Mediator/Test/Test/Classes/Private/Application/TestWebViewController.m
@@ -11,7 +11,7 @@
 // 如果需要隐藏导航栏，可以加载时显示导航栏，WebView延伸到导航栏下面，加载完成时隐藏导航栏即可
 @interface TestWebViewController () <UIScrollViewDelegate>
 
-@property (nonatomic, assign) BOOL isExtendedBottom;
+@property (nonatomic, assign) BOOL toolbarHidden;
 
 @end
 
@@ -32,18 +32,9 @@
         return nil;
     } else {
         return @[
-            [UIBarButtonItem fw_itemWithObject:FWIcon.backImage target:self action:@selector(onWebBack)],
+            FWIcon.backImage,
             FWIcon.closeImage,
         ];
-    }
-}
-
-- (void)onWebBack
-{
-    if (self.webView.canGoBack) {
-        [self.webView goBack];
-    } else {
-        [self fw_closeViewControllerAnimated:YES];
     }
 }
 
@@ -53,37 +44,30 @@
     
     // 底部延伸时设置scrollView边距自适应，无需处理frame
     self.webView.scrollView.contentInsetAdjustmentBehavior = UIScrollViewContentInsetAdjustmentAutomatic;
-    self.isExtendedBottom = [@[@YES, @NO].fw_randomObject fw_safeBool];
-    if (self.isExtendedBottom) {
-        self.edgesForExtendedLayout = Theme.isBarTranslucent ? UIRectEdgeAll : UIRectEdgeBottom;
-    // 底部不延伸时如果显示工具栏，且hidesBottomBarWhenPushed为YES，工具栏顶部会显示空白，需处理frame
-    } else {
-        self.edgesForExtendedLayout = Theme.isBarTranslucent ? UIRectEdgeTop : UIRectEdgeNone;
-    }
+    self.edgesForExtendedLayout = Theme.isBarTranslucent ? UIRectEdgeAll : UIRectEdgeBottom;
+    self.toolbarHidden = YES;
     
     [self renderToolbar];
     [self loadRequestUrl];
 }
 
-- (void)viewDidLayoutSubviews
+- (void)viewWillAppear:(BOOL)animated
 {
-    [super viewDidLayoutSubviews];
-    if (self.isExtendedBottom || !self.fw_isLoaded) return;
-    
-    // 顶部延伸时，不需要减顶部栏高度
-    CGFloat topHeight = (self.edgesForExtendedLayout & UIRectEdgeTop) ? 0 : self.fw_topBarHeight;
-    self.view.fw_height = FWScreenHeight - topHeight - self.fw_bottomBarHeight;
+    [super viewWillAppear:animated];
+    self.navigationController.toolbarHidden = self.toolbarHidden;
 }
 
 - (void)viewWillDisappear:(BOOL)animated
 {
     [super viewWillDisappear:animated];
+    self.toolbarHidden = self.navigationController.toolbarHidden;
     self.navigationController.toolbarHidden = YES;
 }
 
 - (void)renderWebView
 {
     self.view.backgroundColor = [Theme tableColor];
+    self.webView.allowsUniversalLinks = YES;
     self.webView.scrollView.delegate = self;
     self.webView.scrollView.showsVerticalScrollIndicator = NO;
     self.webView.scrollView.showsHorizontalScrollIndicator = NO;
@@ -190,38 +174,14 @@
     }];
 }
 
-- (void)webView:(WKWebView *)webView decidePolicyForNavigationAction:(WKNavigationAction *)navigationAction decisionHandler:(void (^)(WKNavigationActionPolicy))decisionHandler
+- (BOOL)webViewShouldLoad:(WKNavigationAction *)navigationAction
 {
-    if ([self respondsToSelector:@selector(webViewShouldLoad:)] &&
-        ![self webViewShouldLoad:navigationAction]) {
-        decisionHandler(WKNavigationActionPolicyCancel);
-        return;
-    }
-    
-    if ([UIApplication fw_isSystemURL:navigationAction.request.URL]) {
-        [UIApplication fw_openURL:navigationAction.request.URL];
-        decisionHandler(WKNavigationActionPolicyCancel);
-        return;
-    }
-    
     if ([navigationAction.request.URL.scheme isEqualToString:@"app"]) {
         [FWRouter openURL:navigationAction.request.URL.absoluteString];
-        decisionHandler(WKNavigationActionPolicyCancel);
-        return;
+        return NO;
     }
     
-    if ([navigationAction.request.URL.scheme isEqualToString:@"https"]) {
-        [UIApplication fw_openUniversalLinks:navigationAction.request.URL completionHandler:^(BOOL success) {
-            if (success) {
-                decisionHandler(WKNavigationActionPolicyCancel);
-            } else {
-                decisionHandler(WKNavigationActionPolicyAllow);
-            }
-        }];
-        return;
-    }
-    
-    decisionHandler(WKNavigationActionPolicyAllow);
+    return YES;
 }
 
 @end