+++ FWApplication/Classes/Compatible/Toolkit/UIKit/UITableView+Toolkit.swift
@@ -357,22 +357,6 @@ extension Wrapper where Base: UITableView {
 
 extension Wrapper where Base: UITableViewCell {
     
-    /// 设置分割线内边距，iOS8+默认15.f，设为UIEdgeInsetsZero可去掉
-    public var separatorInset: UIEdgeInsets {
-        get { return base.__fw_separatorInset }
-        set { base.__fw_separatorInset = newValue }
-    }
-
-    /// 获取当前所属tableView
-    public weak var tableView: UITableView? {
-        return base.__fw_tableView
-    }
-
-    /// 获取当前显示indexPath
-    public var indexPath: IndexPath? {
-        return base.__fw_indexPath
-    }
-    
     /// 延迟加载背景视图，处理section圆角、阴影等。会自动设置backgroundView
     public var backgroundView: TableViewCellBackgroundView {
         return base.__fw_backgroundView