+++ FWApplication/Classes/FWApplication/Module/Controller/FWWebView.m
@@ -23,7 +23,8 @@
 
 - (void)webView:(WKWebView *)webView decidePolicyForNavigationAction:(WKNavigationAction *)navigationAction decisionHandler:(void (^)(WKNavigationActionPolicy))decisionHandler
 {
-    if ([webView isKindOfClass:[FWWebView class]] && ((FWWebView *)webView).cookieEnabled &&
+    if ([webView isKindOfClass:[FWWebView class]] &&
+        ((FWWebView *)webView).cookieEnabled &&
         [navigationAction.request isKindOfClass:NSMutableURLRequest.class]) {
         [FWWebViewCookieManager syncRequestCookie:(NSMutableURLRequest *)navigationAction.request];
     }
@@ -38,17 +39,29 @@
         decisionHandler(WKNavigationActionPolicyCancel);
         return;
     }
+    
     if ([UIApplication fw_isSystemURL:navigationAction.request.URL]) {
         [UIApplication fw_openURL:navigationAction.request.URL];
         decisionHandler(WKNavigationActionPolicyCancel);
         return;
     }
+    
+    if ([webView isKindOfClass:[FWWebView class]] &&
+        ((FWWebView *)webView).allowsUniversalLinks &&
+        [navigationAction.request.URL.scheme isEqualToString:@"https"]) {
+        [UIApplication fw_openUniversalLinks:navigationAction.request.URL completionHandler:^(BOOL success) {
+            decisionHandler(success ? WKNavigationActionPolicyCancel : WKNavigationActionPolicyAllow);
+        }];
+        return;
+    }
+    
     decisionHandler(WKNavigationActionPolicyAllow);
 }
 
 - (void)webView:(WKWebView *)webView decidePolicyForNavigationResponse:(WKNavigationResponse *)navigationResponse decisionHandler:(void (^)(WKNavigationResponsePolicy))decisionHandler
 {
-    if ([webView isKindOfClass:[FWWebView class]] && ((FWWebView *)webView).cookieEnabled) {
+    if ([webView isKindOfClass:[FWWebView class]] &&
+        ((FWWebView *)webView).cookieEnabled) {
         [FWWebViewCookieManager copyWebViewCookie:webView completion:nil];
     }
     
@@ -719,6 +732,57 @@ static int logMaxLength = 500;
     [_base sendData:data responseCallback:responseCallback handlerName:handlerName];
 }
 
+- (void)registerClass:(id)clazz package:(NSString *)package withMapper:(NSDictionary<NSString *,NSString *> * (^)(NSArray<NSString *> *))mapper
+{
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Warc-performSelector-leaks"
+    NSDictionary<NSString *,NSString *> *bridges = [self bridgeClass:clazz withMapper:mapper];
+    [bridges enumerateKeysAndObjectsUsingBlock:^(NSString * _Nonnull key, NSString * _Nonnull obj, BOOL * _Nonnull stop) {
+        NSString *name = package.length > 0 ? [package stringByAppendingString:key] : key;
+        [self registerHandler:name handler:^(id  _Nonnull data, FWJsBridgeResponseCallback  _Nonnull responseCallback) {
+            [clazz performSelector:NSSelectorFromString(obj) withObject:data withObject:responseCallback];
+        }];
+    }];
+#pragma clang diagnostic pop
+}
+
+- (void)unregisterClass:(id)clazz package:(NSString *)package withMapper:(NSDictionary<NSString *,NSString *> * (^)(NSArray<NSString *> *))mapper
+{
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Warc-performSelector-leaks"
+    NSDictionary<NSString *,NSString *> *bridges = [self bridgeClass:clazz withMapper:mapper];
+    [bridges enumerateKeysAndObjectsUsingBlock:^(NSString * _Nonnull key, NSString * _Nonnull obj, BOOL * _Nonnull stop) {
+        NSString *name = package.length > 0 ? [package stringByAppendingString:key] : key;
+        [self removeHandler:name];
+    }];
+#pragma clang diagnostic pop
+}
+
+- (NSDictionary<NSString *,NSString *> *)bridgeClass:(id)clazz withMapper:(NSDictionary<NSString *,NSString *> * (^)(NSArray<NSString *> *))mapper
+{
+    Class metaClass;
+    if (object_isClass(clazz)) {
+        metaClass = objc_getMetaClass(NSStringFromClass(clazz).UTF8String);
+    } else {
+        metaClass = object_getClass(clazz);
+    }
+    if (!metaClass) return @{};
+    
+    NSArray<NSString *> *methods = [NSObject fw_classMethods:metaClass superclass:NO];
+    if (mapper) {
+        return mapper(methods);
+    }
+    
+    NSMutableDictionary *bridges = [NSMutableDictionary dictionary];
+    for (NSString *method in methods) {
+        if (![method hasSuffix:@":callback:"]) continue;
+        
+        NSString *name = [method stringByReplacingOccurrencesOfString:@":callback:" withString:@""];
+        bridges[name] = method;
+    }
+    return bridges;
+}
+
 - (void)registerHandler:(NSString *)handlerName handler:(FWJsBridgeHandler)handler {
     _base.messageHandlers[handlerName] = [handler copy];
 }
@@ -891,6 +955,7 @@ NSString * FWWebViewJsBridge_js() {
     }
     window.WebViewJavascriptBridge = {
         registerHandler: registerHandler,
+        removeHandler: removeHandler,
         getRegisteredHandlers: getRegisteredHandlers,
         setErrorHandler: setErrorHandler,
         setFilterHandler: setFilterHandler,
@@ -915,6 +980,10 @@ NSString * FWWebViewJsBridge_js() {
 
     function registerHandler(handlerName, handler) {
         messageHandlers[handlerName] = handler;
+    }
+    
+    function removeHandler(handlerName) {
+        delete messageHandlers[handlerName];
     }
     
     function getRegisteredHandlers() {