+++ FWApplication/Classes/FWApplication/Toolkit/UIKit/UITableView+FWApplication.h
@@ -45,19 +45,6 @@ NS_ASSUME_NONNULL_BEGIN
 
 @end
 
-@interface UITableViewCell (FWApplication)
-
-/// 设置分割线内边距，iOS8+默认15.f，设为UIEdgeInsetsZero可去掉
-@property (nonatomic, assign) UIEdgeInsets fw_separatorInset NS_REFINED_FOR_SWIFT;
-
-/// 获取当前所属tableView
-@property (nonatomic, weak, readonly, nullable) UITableView *fw_tableView NS_REFINED_FOR_SWIFT;
-
-/// 获取当前显示indexPath
-@property (nonatomic, readonly, nullable) NSIndexPath *fw_indexPath NS_REFINED_FOR_SWIFT;
-
-@end
-
 /**
  TableViewCell背景视图，处理section圆角、阴影等
  */