+++ docs/docsets/FWApplication.docset/Contents/Resources/Documents/FWApplication/NSAttributedString+FWApplication.h
@@ -109,14 +109,6 @@ NS_SWIFT_NAME(AttributedOption)
  */
 @interface NSAttributedString (FWApplication)
 
-#pragma mark - Convert
-
-/// 快速创建NSAttributedString，自定义字体
-+ (instancetype)fw_attributedString:(NSString *)string withFont:(nullable UIFont *)font NS_REFINED_FOR_SWIFT;
-
-/// 快速创建NSAttributedString，自定义字体和颜色
-+ (instancetype)fw_attributedString:(NSString *)string withFont:(nullable UIFont *)font textColor:(nullable UIColor *)textColor NS_REFINED_FOR_SWIFT;
-
 #pragma mark - Html
 
 /// html字符串转换为NSAttributedString对象，可设置默认系统字体和颜色(附加CSS方式)